Verify OpenSSL tarball checksums
It's not nice to just download whatever source code from the Internet
and build a cryptography library from that. At least verify that the
checksum of the file is the expected one.

The checksums are retrieved via a trusted Web browser from a trusted
network from the official Web site. We check that the file we download
is the one we should expect, just in case openssl.org gets MITMed.