CocoaPods spec: CLOpenSSL (#3)
Support publishing our OpenSSL build on CocoaPods. Now running
"make specs" will also generate "cocoapods/CLOpenSSL.podspec" which can
be uploaded to CocoaPods repositories.

The spec does not actually build OpenSSL on the user machine. Instead,
it downloads prebuilt frameworks (the same as used by Carthage). This
results in considerable savings of the installation time. The framework
binaries are currently hosted as GitHub releases. Not checking them into
the repository also gives some savings (or rather, will give over time).

Since the spec still clones *this* repository, it will be possible to
build OpenSSL from source on developers' machines, if needed.

Only dynamic frameworks are supported with CocoaPods. There is an option
to include static frameworks, but that does not work out of the box and
is very inconvenient for us to support.

Note that at the moment there is no "v1.0.221" tag in this repository so
the produced Podspec will not pass linting. It can be properly and fully
linted only after the tag is published. (That's how Cocoapods works.)