Apple Silicon support (#9)
* Apple Silicon support

Add architecture configuration for OpenSSL based on changes suggested
by Sasmito [1]. Since we're interested only in macOS and iOS, add only
those simulators, cross-compiled for arm64.

This only lets the compiler to compile something as support for arm64
is not complete in OpenSSL [2]. For example, you can see that we disable
inline assembly, so the code might actually run slower. It's not much of
an issue right now though since there is no arm64 hardware yet on the
public market.

[1]: https://github.com/balthisar/openssl-xcframeworks/pull/6
[2]: https://github.com/openssl/openssl/issues/12254

However, don't add the new arm64 architecture to the list of default
architectures and targets to compile OpenSSL for (DEFAULTARCHS and
DEFAULTTARGETS in "build-libssl.h"). This is because arm64 is currently
supported only in Xcode betas, it's not available yet for the stable
releases. Those wishing to experiment with arm64 architecture will have
to specify appropriate targets manually on the command line.

Suggested-by: Sasmito Adibowo <adib@basil-salad.com>

* Differentiate iOS Simulator headers

New arm64 architecture for iOS Simulator is conflicting with arm64
architecture for the real iOS devices. Though the architecture is the
same, the target is not. A different SDK is used to build arm64 for
Apple Silicon devices, and it's not compatible in the strict sense.

OpenSSL currently has per-architecture headers with platform-specific
details. Make sure that iOS ones don't conflict with iOS Simulator by
making iOS use "openssl_ios_arm64.h" and iOS Simulator will be using
"openssl_ios_sim_arm64.h"

* Add platform-specific headers for arm64

Introduce conditionals to check for and use arm64 on macOS and for the
iOS Simulator builds.