Fix minimum OS version setup (#14)
Make sure to export the MIN_IOS_SDK and MIN_OSX_SDK variables from the
makefile as they are needed in "scripts/update-specs.sh" to properly
fill in the minimum version requirements in generated CocoaPods spec.
Otherwise we only pass them to the OpenSSL build scripts and end up
with the spec requesting more broad minimum versions than the range
actually supported by the binaries.

Instead of exporting the variables to specific script invocations, just
make sure that Make exports them to all tools that it launches. That way
it's harder to forget to add them for some script.