XCFramework support and packaging (#21)
* Drop support for OpenSSL 1.0.x

Upstream commit 184673f65097ff127cbac3da8661c5648f13f309 has removed
support for OpenSSL 1.0.x, now the archirectures to build for are
specified only via "--targets" option, "--archs" is not supported.
Do not set it up and do not pass it to the build script.

* Restore removed iOS 10.0 targets

Upstream has moved to using iOS 12.0 as the default minimum supported
version. With this, several architectures has been removed in commit
2be1214e3b1f87478e531c82056a4d9f3c895dc2. However, we are still using
iOS 10.0 as our minimum supported version and still the removed 32-bit
architectures: armv7 and arm7s for iOS devices and i386 for simulators.
If they are not included in the binary frameworks, Xcode will complain
about missing architecture slices.

* Adjust packaging to XCFrameworks

Upstream has moved to packaging a single XCFramework with all targets in
it. This brings good news as now Apple Silicon will be supported for iOS
Simulators, removing the need for some hacks. This also brings bad news
because this is a breaking change for CLOpenSSL users.

In order to avoid breakage, the packages are renamed. CocoaPods' pod is
now named "CLOpenSSL-XCF" and Carthage binary dependencies should be
retrieved using the "openssl-{static,dynamic}-xcframework.json" specs.

The reason for this is that Carthage users will have to adjust their
binary dependencies to download the unified XCFramework instead of two
separate frameworks for iOS and macOS. They will also need to adjust
their Xcode projects to use "openssl.xcframework" instead of the usual
"openssl.framework". While CocoaPods should automatically figure this
out, we won't take chances and will make this a flag day for CocoaPods
users as well.

This is also a breaking change for our build system as the upstream
build system now builds iOS (device) and iOS Simulator frameworks
separately. This means that there are two "openssl.framework" binaries
for iOS and neither is usable with both iOS and iOS Simulator, as it was
before. We can't package them as an "iPhone" target to maintain backward
compatibility with previous Carthage packaging.

Overall, this means that the next release of CLOpenSSL is a flag day for
its users. Users of the previous version MUST migrate their builds if
they wish to future updates. Otherwise they can continue using 1.1.10802.

* Include arm64 slice for iOS Simulators

Now that we are building XCFrameworks, we can also include the arm64
slice for iOS Simulator (if Xcode supports Apple Silicon), without
getting any errors from linker.

It is also not necessary now to exclude this slice from projects
generated by CocoaPods. That is, for CLOpenSSL consumers. Upstream
library users like Themis might still need the exclusion until they are
producing XCFramework-compatible artifacts. However, CLOpenSSL does not
need to enforce this exclusion by itself from now on.